support arrays for names, move support funcs to fhir model, unit tests
