+++ portal/models/fhir.py
@@ -1,5 +1,6 @@
 """Model classes for retaining FHIR data"""
 from datetime import datetime
+from flask import current_app
 from html.parser import HTMLParser
 import json
 from sqlalchemy import UniqueConstraint, or_
@@ -674,3 +675,33 @@ def add_static_concepts(only_quick=False):
     for concept in LocaleConstants(): pass # looping is adequate
     for concept in TxStartedConstants(): pass  # looping is adequate
     for concept in TxNotStartedConstants(): pass  # looping is adequate
+
+
+def v_or_n(value):
+    """Return None unless the value contains data"""
+    return value.rstrip() if value else None
+
+
+def v_or_first(value, field_name):
+    """Return desired from list or scalar value
+
+    :param value: the raw data, may be a single value (directly
+     returned) or a list from which the first element will be returned
+    :param field_name: used in error text when multiple values
+     are found for a constrained item.
+
+    Some fields, such as `name` were assumed to always be a single
+    dictionary containing single values, whereas the FHIR spec
+    defines them to support 0..* meaning we must handle a list.
+
+    NB - as the datamodel still only expects one, a 400 will be
+    raised if given multiple values, using the `field_name` in the text.
+
+    """
+    if isinstance(value, (tuple, list)):
+        if len(value) > 1:
+            msg = "Can't handle multiple values for `{}`".format(field_name)
+            current_app.logger.warn(msg)
+            abort(400, msg)
+        return value[0]
+    return value
